requirements.txt
streamlit==1.22.0
requests==2.28.2
python-dotenv==1.0.0
pillow==9.5.0
langchain==0.0.167
openai==0.27.6
